Moved EnvironmentInformedActionPanel and updated imports to accommodate new lobby/lab model
