+++ ui/src/app/components/runtime-configuration-panel/customize-panel.tsx
@@ -13,12 +13,12 @@ import {
 
 import { UIAppType } from 'app/components/apps-panel/utils';
 import { LinkButton } from 'app/components/buttons';
-import { EnvironmentInformedActionPanel } from 'app/components/common-env-conf-panels/environment-informed-action-panel';
 import { MachineSelector } from 'app/components/common-env-conf-panels/machine-selector';
 import { styles } from 'app/components/common-env-conf-panels/styles';
 import { FlexColumn, FlexRow } from 'app/components/flex';
 import { ErrorMessage, WarningMessage } from 'app/components/messages';
 import { TooltipTrigger } from 'app/components/popups';
+import { EnvironmentInformedActionPanel } from 'app/lab/components/common-env-conf-panels/environment-informed-action-panel';
 import {
   AnalysisConfig,
   maybeWithExistingDiskName,
